+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 * Support environment variable names prefixed with `rp_` in `reportportal:start_launch` and `reportportal:finish_launch` for consistency with formatters
 * Support providing configuration values via upcased environment variables (e.g. `RP_UUID`)
 * Report real, not mocked time when Timecop is used
+* Add commands to modify/search/delete test items